PSMDB-139 Take out compaction-related code from RocksEngine
Put it to RocksCompactionScheduler.
Mostly no functional changes, the only place is in dropIdent where
_identMap is now cleaned up after all other code is executed, but
it's not a big deal.
